requirements.txt
Pillow==9.1.1
uvicorn==0.17.6
fastapi==0.78.0
pydantic==1.9.1
python-multipart